Adpating Qr Code classes
Making changes so that the code can be used for both scanning ticket and adding a friend, quite complex and coneptual but it's making progress
