+++ app/src/main/java/com/github/se/eventradar/ui/qrCode/QrCodeScanCamera.kt
@@ -26,66 +26,74 @@ import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.core.content.ContextCompat
-import com.github.se.eventradar.qrCode.QrCodeScanConstraints
+import androidx.lifecycle.lifecycleScope
+import com.github.se.eventradar.qrCode.QrCodeAnalyser
 
-class QrCodeScanCamera(private val onQrCodeScanned: (String) -> Unit) {
+class QrCodeScanCamera() {
 
-  @Composable
-  fun QrCodeScanner() {
+    @Composable
+     fun QrCodeScanner(friendOrTicket: Int) {
 
-    val context = LocalContext.current
-    val lifeCycleOwner = LocalLifecycleOwner.current
-    val cameraFutureProvider = remember { ProcessCameraProvider.getInstance(context) }
+        val context = LocalContext.current
+        val lifeCycleOwner = LocalLifecycleOwner.current
+        val currentCoroutineScope = lifeCycleOwner.lifecycleScope
+        val cameraFutureProvider = remember { ProcessCameraProvider.getInstance(context) }
 
-    var hasCameraPermission by remember {
-      mutableStateOf(
-          ContextCompat.checkSelfPermission(context, android.Manifest.permission.CAMERA) ==
-              PackageManager.PERMISSION_GRANTED)
-    }
-
-    val launcher =
-        rememberLauncherForActivityResult(
-            contract = ActivityResultContracts.RequestPermission(),
-            onResult = { granted -> hasCameraPermission = granted })
-    // only ask permission once
-    LaunchedEffect(key1 = true) { launcher.launch(android.Manifest.permission.CAMERA) }
-    Column(modifier = Modifier.fillMaxSize()) {
-      if (hasCameraPermission) {
-        Spacer(modifier = Modifier.height(80.dp))
-        AndroidView( // PreviewView !E for Composable hence need ot create AndroidView
-            factory = { context ->
-              val previewView = PreviewView(context)
-              val preview = Preview.Builder().build()
-              val selector =
-                  CameraSelector.Builder()
-                      .requireLensFacing(CameraSelector.LENS_FACING_BACK)
-                      .build()
-              preview.setSurfaceProvider(previewView.surfaceProvider)
-              val imageAnalysis =
-                  ImageAnalysis.Builder()
-                      .setBackpressureStrategy(ImageAnalysis.STRATEGY_KEEP_ONLY_LATEST)
-                      .build()
-              imageAnalysis.setAnalyzer(
-                  ContextCompat.getMainExecutor(context), QrCodeScanConstraints(onQrCodeScanned))
-              try {
-                cameraFutureProvider
-                    .get()
-                    .bindToLifecycle(
-                        lifeCycleOwner, // only launched during current compose lifecycle
-                        selector,
-                        preview,
-                        imageAnalysis)
-              } catch (e: Exception) {
-                e.printStackTrace()
-              }
-              previewView
-            },
-            modifier = Modifier.weight(1.5f).aspectRatio(1f).padding(horizontal = 32.dp)
-            //                Text(
-
-            //                    text = code, )
+        var hasCameraPermission by remember {
+            mutableStateOf(
+                ContextCompat.checkSelfPermission(context, android.Manifest.permission.CAMERA) ==
+                        PackageManager.PERMISSION_GRANTED
             )
-      }
+        }
+
+        val launcher =
+            rememberLauncherForActivityResult(
+                contract = ActivityResultContracts.RequestPermission(),
+                onResult = { granted -> hasCameraPermission = granted })
+        // only ask permission once
+        LaunchedEffect(key1 = true) { launcher.launch(android.Manifest.permission.CAMERA) }
+        Column(modifier = Modifier.fillMaxSize()) {
+            if (hasCameraPermission) {
+                Spacer(modifier = Modifier.height(80.dp))
+                AndroidView( // PreviewView !E for Composable hence need ot create AndroidView
+                    factory = { context ->
+                        val previewView = PreviewView(context)
+                        val preview = Preview.Builder().build()
+                        val selector =
+                            CameraSelector.Builder()
+                                .requireLensFacing(CameraSelector.LENS_FACING_BACK)
+                                .build()
+                        preview.setSurfaceProvider(previewView.surfaceProvider)
+                        val imageAnalysis =
+                            ImageAnalysis.Builder()
+                                .setBackpressureStrategy(ImageAnalysis.STRATEGY_KEEP_ONLY_LATEST)
+                                .build()
+                        imageAnalysis.setAnalyzer(
+                            ContextCompat.getMainExecutor(context),
+                            QrCodeAnalyser(currentCoroutineScope, friendOrTicket)
+
+                        )
+                        try {
+                            cameraFutureProvider
+                                .get()
+                                .bindToLifecycle(
+                                    lifeCycleOwner, // only launched during current compose lifecycle
+                                    selector,
+                                    preview,
+                                    imageAnalysis
+                                )
+                        } catch (e: Exception) {
+                            e.printStackTrace()
+                        }
+                        previewView
+                    },
+                    modifier = Modifier.weight(1.5f).aspectRatio(1f).padding(horizontal = 32.dp)
+                    //                Text(
+                    //                    text = code, )
+                )
+            }
+        }
+
+
     }
-  }
 }
